Add files via upload
The above codes are the implementations of the numerical renormalization group (NRG) procedure which can be applied to quantum impurity systems. This nonperturbative numerical method provides an accurate description of physical properties in all parameter regimes of interest to us in equilibrium. In such systems, the coupling of a small quantum system with a few degrees of freedom to the environment, consisting of a continuum of states, leads to a variety of novel properties. Interactions, such as the local Coulomb repulsion or the local electron-phonon interaction in the small system result in scattering processes that can drastically change the spectral and transport properties revealing fascinating correlation effects. Following the general strategy of NRG, the low-lying many-body eigenstates and eigenenergies of the system can be obtained, which enables us to calculate thermodynamical properties as well as the dynamical quantities such as the spectral function at zero or finite temperatures.